Fix combatTeleportDelay being an int.

+++ src/main/java/net/sudologic/rivals/EventManager.java
@@ -34,9 +34,9 @@ public class EventManager implements Listener {
         deathPowerLoss = (double) settings.get("deathPowerLoss");
         tradePower = (double) settings.get("tradePower");
         if(settings.contains("combatTeleportDelay")) {
-            combatTeleportDelay = (double) settings.get("combatTeleportDelay");
+            combatTeleportDelay = (int) settings.get("combatTeleportDelay");
         } else {
-            settings.set("combatTeleportDelay", 120);
+            settings.set("combatTeleportDelay", 120.0);
         }
         combatTime = new HashMap<>();
     }
